ECO/requirements.txt
dash==2.0.0
dash-bootstrap-components==1.0.0
pandas==1.3.3
numpy==1.21.2
werkzeug==2.0.3
minio
boto3
pyarrow
python-dotenv